Move `@SerialDescription` from annotations to generator-json module
The annotation requires `@SerialInfo` from kotlinx.serialization, which
forces an `api` dependency on Native targets. Since the annotation is only
meaningful with the serialization-based schema generator, it belongs in the
`kotlinx-schema-generator-json` module (`kotlinx.schema.generator.json` package).

- Remove kotlinx.serialization.core dependency from annotations module
- Remove `SerialDescribedProduct` and its KSP integration test
- Update API dumps, imports, and documentation
