+++ kotlinx-schema-annotations/gradle.properties
@@ -1,1 +1,2 @@
 description=KotlinX Schema Common Annotations
+kotlin.suppressGradlePluginWarnings=IncorrectCompileOnlyDependencyWarning
