+++ kotlinx-schema-generator-json/api/kotlinx-schema-generator-json.api
@@ -91,6 +91,15 @@ public final class kotlinx/schema/generator/json/ReflectionFunctionCallingSchema
 	public final fun getDefault ()Lkotlinx/schema/generator/json/ReflectionFunctionCallingSchemaGenerator;
 }
 
+public abstract interface annotation class kotlinx/schema/generator/json/SerialDescription : java/lang/annotation/Annotation {
+	public abstract fun value ()Ljava/lang/String;
+}
+
+public final synthetic class kotlinx/schema/generator/json/SerialDescription$Impl : kotlinx/schema/generator/json/SerialDescription {
+	public fun <init> (Ljava/lang/String;)V
+	public final synthetic fun value ()Ljava/lang/String;
+}
+
 public final class kotlinx/schema/generator/json/TypeGraphToFunctionCallingSchemaTransformer : kotlinx/schema/generator/core/ir/AbstractTypeGraphTransformer {
 	public fun <init> ()V
 	public fun <init> (Lkotlinx/schema/generator/json/FunctionCallingSchemaConfig;)V
