+++ kotlinx-schema-generator-json/src/jvmTest/kotlin/kotlinx/schema/generator/json/JsonSchemaGeneratorTest.kt
@@ -4,7 +4,6 @@ package kotlinx.schema.generator.json
 
 import io.kotest.assertions.json.shouldEqualJson
 import kotlinx.schema.Description
-import kotlinx.schema.SerialDescription
 import kotlinx.serialization.json.Json
 import kotlin.test.Test
